Refactor view package into CLI and GUI packages
