+++ src/main/java/duke/command/DeleteTaskCommand.java
@@ -3,7 +3,7 @@ import duke.command.exceptions.CommandExecutionError;
 import duke.interfaces.Command;
 import duke.model.Task;
 import duke.model.TaskModel;
-import duke.view.TaskView;
+import duke.view.cli.TaskView;
 
 /**
  * Command that deletes tasks.
