+++ src/main/java/duke/command/MarkUndoneCommand.java
@@ -2,7 +2,7 @@ package duke.command;
 import duke.command.exceptions.CommandExecutionError;
 import duke.interfaces.Command;
 import duke.model.TaskModel;
-import duke.view.TaskView;
+import duke.view.cli.TaskView;
 
 /**
  * Command to mark a task as undone.
